the first step , a main method to show an empty GEF canvas zone
In first stage we will  create a simple Java class containing a main(…)
method.

For creating a diagram as part of a larger Eclipse RCP application, then
create a Plug-in project with the following plug-in dependencies:
• org.eclipse.ui
• org.eclipse.core.runtime
• org.eclipse.draw2d

the class contains also two other methods

run method which is responsible to create a shell and fixing it size
then making its content as an empty Canvas. finally displaying it on
screen.

the creatDiagram method will add the possibility to make different
graphics in the canvas. initially it will add an empty figure as hole
container.