adding connections
to connect each of the people to the marriage (see Figure 2–5 page 14
GEF book ), showing their relationship to one another by modifying the
createDiagram(…) method again. This is accomplished by calling a
connect(…) method to create the line connecting the center of one figure
to the center of another (see Chapter 6 on page 69 for more about
connections).