Adding Moving Figures Capability
When the user presses the mouse button, we need to record the location
where the mouse down occurred by adding a field and implementing the
mousePressed(…) method. In addition, this method must mark the event as
“consumed” so that the Draw2D event dispatcher will send all mouse
events to this listener’s figure until the mouse button is released.

As the user moves the mouse around with the mouse button held down,
we need to move the figure in the same direction and distance. The
mouseDragged(…) method calculates the distance moved, moves the figure,
and marks the event as consumed. To move the figure, we must update both
the figure’s bounding box and the layout information. Both the figure’s
original location and new location must be marked as “dirty” so that the
update manager will redraw the diagram appropriately. The getBounds()
method returns the actual rectangle object used by the figure to
remember its bounds, so we cannot modify that object. Instead, we call
getCopy() before calling translate(…) to prevent any undesired side
effects.